requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
brotlipy=0.7.0=py38h94c058a_1001
bzip2=1.0.8=hc929b4f_4
c-ares=1.17.1=hc929b4f_0
ca-certificates=2020.12.8=hecd8cb5_0
cairo=1.16.0=h4ec454a_1007
certifi=2020.12.5=py38h50d1736_1
cffi=1.14.4=py38h979bc6b_1
chardet=4.0.0=py38h50d1736_1
click=7.1.2=py_0
cryptography=3.3.1=py38h6b4ec92_0
decorator=4.4.2=py_0
dlib=19.21.1=py38hfe90ab8_0
ffmpeg=4.3.1=hb787a4c_2
flask=1.1.2=py_0
flask-cors=3.0.9=py_0
fontconfig=2.13.1=h07ff6b0_1003
freetype=2.10.4=h3f75d11_0
gettext=0.19.8.1=h7937167_1005
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.13=h2e338ed_1001
harfbuzz=2.7.4=hf28d55f_0
hdf5=1.10.6=nompi_hc5d9132_1113
icu=68.1=h74dc148_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
imageio-ffmpeg=0.4.3=pyhd8ed1ab_0
intel-openmp=2020.2=258
itsdangerous=1.1.0=py_0
jasper=1.900.1=h636a363_1006
jinja2=2.11.2=py_0
jpeg=9d=hbcb3906_0
krb5=1.17.2=h60d9502_0
lame=3.100=h35c211d_1001
lcms2=2.11=h11f7e16_1
libblas=3.9.0=7_openblas
libcblas=3.9.0=7_openblas
libcurl=7.71.1=h9bf37e3_8
libcxx=11.0.0=h4c3b8ed_1
libedit=3.1.20191231=h1de35cc_1
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=h6c81a4c_15
libgfortran5=9.3.0=h6c81a4c_15
libglib=2.66.4=h7424822_1
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=7_openblas
liblapacke=3.9.0=7_openblas
libnghttp2=1.41.0=h7580e61_2
libopenblas=0.3.12=openmp_h54245bb_1
libopencv=4.5.0=py38_6
libopus=1.3.1=hc929b4f_1
libpng=1.6.37=h7cec526_2
libssh2=1.9.0=h8a08a2b_5
libtiff=4.1.0=hcb84e12_0
libvpx=1.7.0=h378b8a2_0
libwebp-base=1.1.0=hbcb3906_3
libxml2=2.9.10=h802ca4f_3
llvm-openmp=11.0.0=h73239a0_1
markupsafe=1.1.1=py38h1de35cc_1
mkl=2020.2=260
moviepy=1.0.1=py_0
munch=2.5.0=pypi_0
ncurses=6.2=h0a44026_1
nettle=3.6=hedd7734_0
ninja=1.10.2=py38hf7b0b51_0
numpy=1.19.5=py38h6ced74f_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py38_6
openh264=2.1.1=hfd3ada9_0
openssl=1.1.1i=h35c211d_0
pcre=8.44=hb1e8313_0
pillow=8.1.0=py38hc1d52f7_0
pip=20.2.4=py38_0
pixman=0.40.0=hbcb3906_0
pretrainedmodels=0.7.4=pypi_0
proglog=0.1.9=py_0
py-opencv=4.5.0=py38h9bb44b7_6
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py38h5347e94_2
python=3.8.5=h26836e1_1
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_0
readline=8.0=h1de35cc_0
requests=2.25.1=pyhd3deb0d_0
setuptools=50.3.0=py38h0dc7051_1
six=1.15.0=py_0
sqlite=3.34.0=h17101e1_0
tk=8.6.10=hb0a8c7a_0
torchaudio=0.4.0=py38
torchvision=0.5.0=py38_cpu
tqdm=4.55.1=pyhd8ed1ab_0
urllib3=1.26.2=pyhd8ed1ab_0
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
x264=1!161.3030=h35c211d_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
